Add golang formatting documentation
This patch is migrated from [0] as the repo used to store Airship
docs has changed to airship/docs.

This patch addresses feedback found in the other patch by removing
the manual import order specified to follow PEP8 standards as found
in the OpenStack docs in favor of using the goimports tool.

It is expected that when code is submitted it should have gofmt and
goimports run against it before hand - but these will be enforced
at the gate.

Change-Id: I624bade922e034205643a5214295a0febea7953b