Remove AIAP documentation fork
This removes the copy of AIAP documentation, and forwards the reader
instead to the AIAP README from which the copy was forked.  This is
needed because the AIAP documentation has continued to evolve,
so the deleted copy is out of date.

Any changes made to the deleted document which are still relevant
will need to be added to the AIAP README.

Change-Id: I39a11b6edd5d4ef893d66a5e7e7be4186f895ace